Create cc::Service and cc::ServiceFactory.
The cc::Service is the root of the ServiceCompositor, and owns things
like the LayerTreeHostImpl and the Scheduler. It will roughly replace
the ProxyImpl.

cc::Service implements gpu::ServiceCompositor.

cc::ServiceFactory implements gpu::ServiceCompositorFactory, and
it returns a new gpu::ServiceCompositor.

This lets content/gpu/ create a cc::ServiceFactory, and pass it as a
gpu::ServiceCompositorFactory over to gpu/ipc/service.

GpuChannel makes a cc::Service (as a gpu::ServiceCompositor) via this
factory when the CreateCompositor IPC is called.