Make OutputSurface and ContextProviders for the service compositor and Display.
This makes the following classes:

- ServiceContextProvider. Uses GLInProcessContext, and a simple
InProcessCommandBuffer::Service implementation to run all tasks on the same
thread (which is the GPU main thread), and is copied/stripped down from
webview (android_webview/browser/deferred_gpu_command_service.cc). The
ServiceContextProvider can be on or offscreen depending on a widget being
present.

- DelegatingOutputSurface. This is a rip-off of SurfaceDisplayOutputSurface
with a few changes. It doesn't expect to share a context with the Display,
and a Display is optional (it's null for the non-root compositors). It may
need to change more in the future to have more things from
DelegatedFrameHost, I'm unclear on this.

- DisplayOutputSurface. This is a child of PixelTestOutputSurface and
GpuBrowserCompositorOutputSurface. It uses a sync token to see when the
swap is complete since we don't have a signal from the
InProcessCommandBuffer available.

The Service creates all of these things. It owns the
DelegatingOutputSurface that is given to the LayerTreeHostImpl. It also
optionally owns a Display (if it has a widget aka is a root compositor),
and creates a DisplayOutputSurface for that Display.

ServiceFactory owns the global SurfaceManager for the system.