Push the tree structure to create LayerImpls. [STILL CRASHES]
This pushes the layers from LTH to LTHI. It does not push any layer
properties/state yet. That needs to be done by the push properties
step, and can use Layer::WritePropertiesMojom, and
LayerImpl::ReadPropertiesMojom.

The patch currently crashes in ComputeVisibleRects, which may be
because of missing properties, or something related to property
trees. It's unclear.