Setup to inform parent compositors about new SurfaceIds from their children.
This is meant to deal with the flashing missing renderer while it is producing
its first frame. It will allow the parent to show solid color until their
child first swaps a frame then replace it with their child's surface.

On resize we could potentially use this too, to lock the parent compositor in
place until the child is done or something.. *hand wave*

Some TODOs to fill in and an IPC needed to inform the parent, and have it act
on it to switch from solid color to surface content.