DisplayCompositorHost skeleton implementation
This change implements a skeleton DisplayCompositorHost. In Chrome,
this would be the browser process. In Mus+Ash, this would be the
window server. The DisplayCompositorHost is provided by a privilieged
process and allows unprivilged processes to gain access to the
compositor. Connections to the compositor flow through the host first
which, in a subsequent change, would allocate a globally unique
client ID. In the event of "display compositor" crash (Gpu process
crashes) then, the host is responsible for bringing up a new display
compositor service and passing a client's request onto the new service.

As a privileged actor, the DisplayCompositorHost will
also receive notification of newly created SurfaceIds, and a master
token for each new surface ID. This master token keeps the surface alive
until it is returned to the display compositor. This again, will be
implemented in a subsequent CL.