Fix SurfaceId Refence Counting
This CL makes progress towards making surface IDs do the right thing.
We were double decrementing surface IDs in some cases. We also had an
incorrect DCHECK, and we weren't reaping Surfaces when the refcount went
to 0 but rather after we Destroyed another surface.

We now garbage collect when refcount goes to zero and refcounting should
be more correct.