Don't release surface IDs immediately
The previous frame may not have been activated and a CompositorFrame may
not have been generated for it yet. If we release a surface ID immediately then
we may garbage collect the surface prior to the previous frame
activating. This may result in flashing during resize. This change fixes
this by plumbing released surfaces through ContentFrames and only
releasing them on the swap.