Use DisplayCompositorHost in the browser as well.
In Mus+Ash, the window server is the DisplayCompositorHost and so
the browser would need to connect to the DisplayCompositorHost first
to grab a connection to the display compositor. SurfaceID management
lives on the "CompositorChannel" interface currently. That interface
can only increment refcount or make a temp refcount non-temp, but it
cannot decrement refcount and it cannot allocate a new temp ref.
A ref can only be decremented on SwapBuffers to make sure that existing
references of the surface ID have made their way through the pipeline.

"CompositorChannel" method calls are not handled in order with
Compositor interfaces, and DisplayCompositor. This is probably
OK because we can only increment surface ID refs out of order, and
all new surface IDs start with a temp ref.