All clients share a single DisplayCompositorConnection
This CL begins the plumbing for notifying the display compositor host
of new surfaces submitted to new surface Ids in the display compositor.

This CL makes all display DisplayCompositorHosts share a single
DisplayCompositorConnection. The DisplayCompositorConnection lives
on the IO thread of the browser process. In Mus+Ash, this would be
the Mus window server process. Thus, even the browser UI has a
mojo connection to the DisplayCompositorHost which relays MessagePipes
to the GPU process.

DisplayCompositorHostProxy is used by the browser's UI thread to relay
messsages to the IO thread / window server and back.