There is now a single DisplayCompositor in the GPU Process.
DisplayCompositor now lives entirely on the compositor thread in the
gpu process and has a SurfaceManager. mojom::CompositorChannel is implemented in
DisplayCompositor. It is not tied to a particular view
(RenderWidgetCompositor/LayerTreeHost). It is effectively a connection to some
SurfaceManager methods and so only one connection needs to exist per
client process and can be reused by multiple clients within that process as it
does not rely on client-specific state. "Service" (needs to be renamed) is
roughly a "ContentFrameSink" (~a two-way connection to a LayerTreeHostImpl).
Each window/display has a gpu::SurfaceHandle. Non-top-level surfaces have the
null handle (gpu::kNullSurfaceHandle).

This CL also plumbs OnSurfaceCreated from SurfaceManager =>
DisplayCompositor => DisplayCompositorConnection (in the display
compositor host). There is a single DisplayCompositorConnection as
there is a single display compositor. New "Service" (ContentFrameSink)
connections are established through the display compositor host so that
the display ocmpositor host can create hierarchical associations between
the clients and manage the lifetime of surfaces (resources in general).

TODO: This all needs pretty pictures in a design doc.