Add a sink_id to SurfaceId.
Client IDs are allocated by the display compositor host (browser/window
server). Local IDs + Nonce change every time we create a new sized
CompositorFrame. What if we want to create a new
"ContentFrameSink/CompositorFrameSink" in the client and immediately
refer to that surface in the embedder? We need another component to
surface ID that uniquely identifies a Sink that's
allocated by the client.