ServiceFactory => DisplayCompositorFactory
cc::ServiceFactory has been renamed to cc::DisplayCompositorFactory.
This clarifies the role of this class. The DisplayCompositorFactory is
an interface requested from the GPU process (or browser in Chrome
today). A DisplayCompositorFactory holds a compositor thread and creates
the ("singleton") DisplayCompositor on the compositor thread. That way
the display compositor doesn't need to PostTask back and forth in every
method. Mojo bindings are bound to the compositor thread. Alternative
versions of the DisplayCompositorFactory implementation might exist that
don't introduce a compositor thread. This might be used as a good
transition step in Chrome today.