DisplayCompositorHost::Delegate => DisplayCompositorConnectionFactory
How we get a DisplayCompositorConnection depends on where the
DisplayCompositor lives: gpu process, browser process, or window server.
DisplayCompositorHost::Delegate served the purpose of delegating that
decision to a layer above cc. I've moved the class out to a separate
file and gave it a more descriptive name to clarify the need for this
class.