Make PictureCache work
When serializing, we build a per-layer cache of every SkPicture id we encounter. We keep that
cache until the next serialization. For every SkPicture we encounter, we look to
see if it was in the old (previous) cache, or the new (in-progress), in which
case we skip serialization.
On the receiving side, we similarly keep an old cache and build a new one as we
deserialize, and keep the deserialized pictures there, so that we can look up
the ones that we haven't serialized.