Merge cc::mojom::SurfaceManager into cc::mojom::DisplayCompositor
The SurfaceManager interface provided a direct connection from the
browser's UI thread to the display compositor in the gpu process.
However this added an extra MessagePipe, and API complexity. In the
refcounting approach for surface lifetime management, we end up with an
extra thread hop every time we increment a ref count but we end up with
a wholistic view of refs on every surface in the display compositor.

This could be useful for a side channel to decrement refs on surfaces
after clients crash. For example, what if a parent crashes? How do we
cleanup the resources of the child? If the parent embedder process crashes,
the display compositor host would immediately know about it and since it
knows what refs it holds, it can clear those refs on its behalf.