Move Surface Client Hierarchy Registration to DisplayCompositor API
Previously, relationships between surface clients was established
through ContentFrameSink. ContentFrameSink is an unprivileged API
which means any client can set its relationship relative to any other
client. This seems a bit broken. Only the display compositor host should
be allowed to set relationships among clients. This change moves that
API to the DisplayCompositor API and DisplayCompositorHostPrivate API
for the browser to better match that expectation. This also addresses a
TODO.