Allow the browser to map back surface IDs to clients
In a previous change, I made surface ID consist of four components:

1. Client ID 2. Sink ID 3. Local ID 4. Nonce

The client ID is allocated by the display compositor host.
The sink ID is allocated by the client for each ContentFrameSink(or
CompositorFrameSink)
The local ID and nonce are updated every size or device scale factor
change.

In this change the client ID and sink ID plumbing is complete.

Furthermore, content::DisplayCompositorConnectionFactoryImpl assigns the
client ID based on the process ID of the client (the browser being 0),
and the routing ID of the client. Currently this is a direct mapping
which isn't secure because we shouldn't be passing process IDs to other
processes. However, the display compositor host now has a way to map
surface IDs passed from the gpu in "OnSurfaceCreated" to things that the
browser understands. From there, the browser can use that info to embed
the new surface ID in the parent client. The "display compositor host"
has been empowered to add and remove refs (not implemented) as
necessary.

If a client process crashes, we also need to be allowed to remove refs
on all surface IDs currently ref'ed. The browser can track all of this
now. The appropriate bookkeeping is, however, TBD.