ServiceConnection => ContentFrameSinkConnection
This change is pure moving and renaming in preparation for
display compositor restartability support.

ServiceCompositor => ContentFrameSink, highlighting the distinction from
DisplayCompositor and DisplayCompositorHost interfaces.

ServiceConnection => ContentFrameSinkConnection

This code has been moved to cc/client.

In a future change, ContentFrameSinkConenction will likely gain extra
functionality. It will detect when a connection is lost, and inform
its client which will trigger a request for a new connection. This will,
in turn, flow through the DisplayCompositorHost interface, which will
detect that the DisplayCompositorConnection is gone and ask the
DisplayCompositorConnectionFactory (which in this prototype lives in
content) to create a new DisplayCompositorConnection. This will create
and initialize a new GPU process.