Introduce Initial Plumbing for Restartable Display Compositor
This code is likely more complex than it needs to be and it's
incomplete.

The display compositor successfully restarts but the existing
LayerTreeHosts don't get BeginMainFrames. New LayerTreeHosts for windows
like menus and dialogs are successfully created and do tick and do
update on screen.

The way this works is when a SimpleProxy is created and wishes to talk
to the ContentFrameSink that isn't available yet, it requests a new
ContentFrameSinkConnection from its LayerTreeHost. In turn,
LayerTreeHost requests a new ContentFrameSinkConnection from its
LayerTreeHostClient. In the renderer, RenderWidgetCompositor is a
LayerTreeHostClient, and so it delegates further (icky :-( ) to
RenderWidget which delegates to RenderThreadImpl. From RenderThreadImpl,
we request a ContentFrameSink from the Display Compositor Host (browser
in Chrome, window server in Mus+Ash). The display compositor host
creates a new gpu process if necessary (via
DisplayCompositorConnectionFactory), and relays the message to the gpu
process.

The browser goes through a similar flow. ui::Compositor is a
LayerTreeHostClient. Thus, LayerTreeHost asks ui::Compositor for a new
ContentFrameSinkConnection. ui::Compositor goes to
GpuProcessTransFactory which then goes to BrowserGpuChannelHostFactory
whcih then talks to the DisplayCompositorHost interface (this lives on
the IO thread), and then the flow is the same.

When a new ContentFrameSinkConnection is passed into SimpleProxy we
reset a bunch of state to ensure that a fresh set of layers and state is
pushed through the ContentFrameSink.

The code in this change isn't quite right however, but I'm pushing it
now to capture my changes sooner rather than later.