Display Compositor Restart Mostly Works
There was a need to resend a full layer list and push properties for
the layers in additon to clearing the picture ID caches in the clients.

We allocate a new Surface Id when the display compositor restarts. The
content area of the browser may flicker if the browser commits before
the renderer.

We drop surface IDs that no longer exist. It is highly unlikely for
surface IDs to be reused across gpu instances because there is a nonce
that is randomly generated.

Switching tabs seems broken now during deserialization. The picture ID
cache in the client does not match the picture ID cache in the
ContentFrameSink and so sadness ensues. I'm not yet sure what the root
cause of the problem is. I can work around the problem by simply
dropping the offending picture in the ContentFrameSink.

Flickering of the renderer also needs to be addressed.