Always request a new BeginMainFrame after display compositor restart
The renderer wasn't always getting a BeginMainFrame after display
compositor restart and so the browser window had reference to a
non-existent surface from the previous display compositor and thus
displayed a corrupted content area. This change makes it so that
clients always request a new BeginMainFrame after getting a new
ContentFrameConnection if the display compositor has rebooted.

TODO: There's still a visibility issue to resolve.