DisplayCompositorConnectionObserver => DisplayCompositorConnectionClient
This patch introduces the beginnings of ContentFrameSinkPrivatePtr which
will be owned by the display compositor host. Currently, the
RegisterClientHierarchy/UnregisterClientHierarchy API is inherently race
prone because it may be used prior to the creation of a
ContentFrameSink. In order to ensure that races are not possible, this
CL is moving us to a stage where the hierarchy registration is tied to
the lifetime of the ContentFrameSink.