Introduce ContentFrameSinkPrivate interface for display compositor host
The display compositor host is responsible for registering surface
hierarchies for BeinFrameSource routing. However, a child surface may
acquire a FrameSinkId prior to its parent. The parent must be careful to
queue the hierarchy registration request until it also has a
FrameSinkId. Furthermore, while currently hierarcies can be registered
before associated FrameSinks are created, this extra complexity should
not be strictly necessary.

In this architecture, as with the Gpu service, all ContentFrameSink
requests go through the display compositor host first. From there, the
display compositor host passes along a message pipe to a
ContentFrameSinkPrivate interface to the ContentFrameSink created in the
GPU. Surface children can be added via this interface. All calls to the
interface before binding to an actual ContentFrameSink will be
automagically queued up by mojo until the ContentFrameSink is available
in the GPU.

This essentially makes hierarchy registration a race-free API.

This CL also introduces a mechanism whereby a display compositor host
object (DelegatedFrameHost, for example) can register both a private
ContentFrameSinkPrivate interface once an associated ContentFrameSink is
created, and a DisplayCompositorConnectionClient which will receive
notification of newly submitted surfaces. I'm not yet convinced that
this API makes sense and so it may end up being the case that I will
delete this.