Introduce new ContentFrameSinkObserver mechanism for the browser
The browser needs to be able to set up the BeginFrame hierarchy in
Chrome. BeginFrame hierarchy needs to be set by from DelegatedFrameHost
as well as RenderWidgetHostViewChildFrame/RenderWidgetHostViewGuest.

DelegatedFrameHost and RWHVCF/RWHVG also needs to be informed of newly
created surfaces associated with their clients in order to update the
parent client's associated SurfaceLayer.

This all suggests that ui::Compositor (for symmetry),
DelegatedFrameHost, and RWHVCF/RWHVG need to be tied to a
ContentFrameSink. In the past, hierarchy registration and updating of
SurfaceLayers was triggered by convention Chrome IPCs passed all the way
from the LayerTreeHosts of the clients. The downside of all of this is
clients can lie about their surfaces, and we need to validate clients by
parents of a Surface ID they can't lie about (the client ID). Clients
could also send us non-existant surface IDs resulting in artifacting on
the screen. It seems cleaner to have the display compositor directly
tell interested end points in the browser about updated surfaces and
allow them to register BeginFrame hierarchies in a race-free way.

This CL takes us in that direction. We no longer worry about
CompositorFrameSinkIds from the clients or SurfaceIds. A new
registration mechanism has been added that allows an endpoint in the
browser to observe surface IDs as they are created from a particular
client, and register hierarchies in a race-free way.

This CL also takes us in a direction where we can re-register
hierarchies in a race-free way even if the display compositor crashes
and is restarted. All CompositorFrameSinkIds are maintained by the
display compositor host, and the browser knows about the hierarchies and
it can immediately create hierarchies even before clients starts
requesting new ContentFrameSinks. When ContentFrameSinks start getting
created, the observers will be attached to the request. If the observer
is attached after the sink is created, then the observer will
immediately be informed of the last surface ID created, and its queued
messages will be passed to the ContentFrameSink endpoint. This ensures
that we always have a valid hierarchy regardless of order of operations.

This is not quite true in this CL just yet though because when the
display compositor goes away, all the endpoints get informed
asynchronously in an arbitrary order, and thus, messages may end up
going to the wrong endpoints. I haven't fully thought through a solution
for this, but it probably involves using associated interfaces to the
display compositor host and PostTask'ing ContentFrameSinkObserver
registration instead of immediately registering.