Introduce SurfaceId Reference Attribution
This change moves AddRefOnSurfaceId to ContentFrameSinkPrivate,
gets rid of the idea of temp refs (because we trust the display
compositor host to do the right thing), and introduces the idea of a
transferred ref.

1. We need attribution of references to surfaces so that if an embedder
goes away, it references are automatically cleaned up. Thus AddRef
should be part of the ContentFrameSinkPrivate interface. Only the
display compositor host (and the display compositor itself) can add
refs. A client can release refs at any time but the ContentFrameSink
will verify that it owns the refs its releasing.

2. Temp refs existed because the gpu was giving the surface ID first to
the child which held a temp ref and passed it onto the browser. Now
Surface IDs arrive directly to the browser (display compositor host) from the
gpu and so we trust that the display compositor host wil behave. Thus
temp refs can go away.

3. Transferring refs is a way to transfer a ref from ownership by the
display compositor host to some other client embedder without doing two
IPCs. This is purely an optimization.