Remove ContentFrameSink::PrepareCommitSync and SurfaceIdAllocator
ContentFrameSink::PrepareCommitSync existed because surface ID
management was designed with the assumption that the GPU always knew
about the Surface ID prior to passing the ID to the browser and
incrementing the ref count on the ID. That assumption is no longer
necessary because surface IDs are passed from the display compositor
(GPU) to the display compositor host (browser) now and thus the browser
only becomes aware of the surface ID AFTER the GPU does. This change
also moves SurfaceIDAllocator to the client, and so new IDs can be
allocated asynchronously.