Make Display Compositor Host manage lifetime of ContentFrameSink
Previously, the lifetime of ContentFrameSink was managed by the client
that requsted the FrameSink. This causes the bookkeeping "AddRef" API
to be racy if the client crashes. In the case of a client crash, the
ContentFrameSink may go away before the updated bookkeeping makes its
way to the ContentFrameSink, and so we may leak surfaces.

This CL shifts ownership to the display compositor host's
ContentFrameSinkObserver object. As long as an observer is registered on a
ContentFrameSink then the ContentFrameSink's state will remain in the
display compositor. The display compositor will inform the observer that
the client connection has been lost. Upon that signal, the
ContentFrameSinkPrivate connection will be closed, causing the
ContentFrameSink to be deleted.

Note that there is still a potential race here that is not breaking but
not ideal: if the display compositor host decides to drop an observer,
then the ContentFrameSink goes away and the client interprets that as a
lost display compositor and may request a new ContentFrameSink. In the
meantime, the client is signaled via another channel to clear its state
(navigation), and thus throw away the ContentFrameSink it just
created.

Another thing to watch out for is what happens if
RegisterContentFrameSinkObserver is never called? Then the
ContentFrameSink will leak. A possible alternative is for the lifetime
of the ContentFrameSink to be the greater of the
ContentFrameSinkObserver and ContentFrameSink client connections. If a
client creates a ContentFrameSink and closes the connection, and there
is no observer installed, then the ContentFrameSink should go away. This
change does not yet implement this. We will probably need all this extra
logic for an upstream version of this code.