Better ContentFrameSink lifetime management
A ContentFrameSink survives as long as EITHER the client or private
connection is live. Once both go away, we destroy the
ContentFrameSink. If the client goes away and there is no observer
registered in the display compositor host, then the display compositor
host will signal the display compositor to get rid of the
ContentFrameSink.